django/to_do_app/requirement.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.3
djangorestframework-jwt==1.11.0
install==1.3.4
Pillow==8.1.2
pkg-resources==0.0.0
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
